+++ CHANGELOG.md
@@ -5,6 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+# [0.0.11] - 2021-09-22
+## Added
+- Made a Milky Way-textured skybox
+
+# [O.0.9] - 2021-09-22
 ## Added
 - Added texture of the Earth to directory
 - Applied texture on the planet